Attempting day 17 with polymorphism :)
For future me trying to solve this:

It is probably not the most time-efficient solution, but one that works
for me.

I'm creating classes for each different shape, all which inherit from
their mother abstract class "Shape." Each shape behaves differently in
terms of their positioning (which is the only real thing needed for make
this work, I think). Then, a Tick function will take in the current
action and move the rock. Further, the Tick function will decide whether
or not to update a surface vector which holds all the points that form
the current surface of the structure being formed. While all of this is
taking place, the maximum height is also being updated.

Hopefully this works...